streamlit/.ipynb_checkpoints/requirements-checkpoint.txt
pandas
psycopg2-binary
seaborn
matplotlib
sqlalchemy
plotly
streamlit
python-dotenv
requests
import-ipynb